Add local-image-proxy command to proxy container image fetches
This command runs a webserver that can fetch manifests and blobs;
it's a bit like an OCI/Docker Distribution API except we can fetch
from multiple registries.  It may be more useful to think of it
as a streaming HTTP version of `podman pull`.

The primary intended use case is for ostree+containers integration.
What we want is to fetch and *stream* the underlying tarball from a container
image that will be unpacked and used for operating system updates.
The ostree code here will do things like validate the embedded
OSTree signatures inside the tarball.

Design constraints/goals:

 - This should not read or write to the containers/image storage; we
   want to stream the result and have other code unpack it
   into the target ostree repo.  I don't want to have to think about
   any potential interactions between the operating system update mechanism/storage
   and container storage.
 - I want to stream instead of pulling the whole image to a temporary
   directory and then re-parse it.
 - But, crucially we *do* want to reuse any configuration in `/etc/containers/registries.conf`
   particularly around mirroring.  In particular for e.g. OpenShift
   it must work to set up Image Content Source Policy to pull from
   a local registry mirror of the CoreOS update stream.
 - We want to reuse the containers/image default handling for
   authentication (`~/.docker/config.json` etc.)
 - I'd really rather not vendor in all of containers/image into
   another program and have to keep it up to date
 - ostree is in C and Rust, adding Go in-process would be a *huge* change
 - By having this out of process, it can easily be runnable fully unprivileged under e.g. a transient
   uid to fully ensure it has no effect on the system and can be sandboxed.

It's of course a bonus here if we also support containers signing, etc.

The expected primary use case is that ostree tooling allocates a `socketpair()`
and passes it down to skopeo, which will run a webserver on that anonymous
socketpair.

For ease of debugging/testing, we also support listening on a localhost TCP port.

More information here: https://github.com/ostreedev/ostree-rs-ext/tree/b84955aa170bf4b5518603547f33d1aec7719f44#pulling-an-ostree-container-directly