Post-pick adjustment
In libpod/image/layerTree.toOCI(), if we can't read the image's
configuration in OCI format, log the error and return a `nil` pointer
instead of returning the error.

2.2's default logging level is "error" instead of "warn" used in the
version the previous commit is backported from.  The error that causes
us to fail in 2.2 is also in a different location, so we need to add a
similar change for it, and update tests to expect the accompanying error
message.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>