[CI:DOCS] Stop conversion of `--` into en dash
In the documentation source `--` is only used in long option names or
console output. Currently any unescaped `--` that's not in a code
block/inline gets converted into a single "en dash". This is confusing
to the reader because long options start with two dashes, not one.

This switches off the dash transformation in Sphinx. Quotation marks,
apostrophes and ellipses are still transformed as before.

It also switches off all smart transforms in the podman-remote windows
documentation build. This matches the Linux man page build, which
doesn't use smart transforms.

Signed-off-by: Rob Cowsill <42620235+rcowsill@users.noreply.github.com>