PoC support for shimv2 runtime to run kata containers v2
This PR is a PoC to integrate the shimv2 as a container runtime to run
kata containers version 2. A very basic functionality has been added to
create, run and stop containers. Some actions are a bit buggy, mainly
those related to attach, interactive and tty but the options can be used.

Signed-off-by: Eduardo Vega <edvegavalerio@gmail.com>