Reimplement multi-arch image build/push workflow
Previously, this workflow was 95% duplicated across Podman, Skopeo and
Buildah repos.  This new workflow addresses this by utilizing a pair of
purpose-built actions to isolate all the per-repo. logic from the
workflow.  This allows the workflow to be simpler/shorter and 100%
duplicated across all three repos - minus the maintenance burden.

Signed-off-by: Chris Evich <cevich@redhat.com>