seccomp: propagate defaultErrnoRet
This fixes the issue of not propagating defaultErrnoRet from
system's seccomp.json to container's config.json, which can
result in redundant seccomp rules and inability to start a container
when using some runtimes (notably runc up to and including 1.0.1).

The fix is to vendor in containers/common with the fixed pkg/seccomp
which no longer ignores defaultErrnoRet.

Brought to you by:
  go get github.com/containers/common@v0.38.18

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>